+++ annotations/models.py
@@ -830,27 +830,35 @@ class RelationTemplatePart(models.Model):
     predicate_concept = models.ForeignKey(Concept, blank=True, null=True,
                                           related_name='used_as_concept_for_predicate')
     predicate_prompt_text = models.BooleanField(default=True,
-    help_text=help_text(
+                                                help_text=help_text(
     """
     Indicates whether the user should be asked for evidence for predicate.
     """))
 
     predicate_description = models.TextField(blank=True, null=True)
 
-    object_node_type = models.CharField(choices=NODE_CHOICES, max_length=2, null=True, blank=True)
+    object_node_type = models.CharField(choices=NODE_CHOICES, max_length=2,
+                                        null=True, blank=True)
     object_label = models.CharField(max_length=100, null=True, blank=True)
-    object_type = models.ForeignKey(Type, blank=True, null=True, related_name='used_as_type_for_object')
-    object_concept = models.ForeignKey(Concept, blank=True, null=True, related_name='used_as_concept_for_object')
-    object_relationtemplate = models.ForeignKey('RelationTemplatePart',  blank=True, null=True, related_name='used_as_object')
+    object_type = models.ForeignKey(Type, blank=True, null=True,
+                                    related_name='used_as_type_for_object')
+    object_concept = models.ForeignKey(Concept, blank=True, null=True,
+                                       related_name='used_as_concept_for_object')
+    object_relationtemplate = models.ForeignKey('RelationTemplatePart',
+                                                blank=True, null=True,
+                                                related_name='used_as_object')
     object_relationtemplate_internal_id = models.IntegerField(default=-1)
-    object_prompt_text = models.BooleanField(default=True)
-    """Indicates whether the user should be asked for evidence for object."""
+    object_prompt_text = models.BooleanField(default=True, help_text=help_text(
+    """Indicates whether the user should be asked for evidence for object."""))
+
     object_description = models.TextField(blank=True, null=True)
 
 
-
-
 class TemporalBounds(models.Model):
+    """
+    DEPRECATED. We now fully implement the Quadruple model in VogonWeb. See
+    :class:`.DateAppellation`\.
+    """
     start = TupleField(blank=True, null=True)
     occur = TupleField(blank=True, null=True)
     end = TupleField(blank=True, null=True)
