+++ annotations/tasks.py
@@ -172,140 +172,7 @@ def save_text_instance(tokenized_content, text_title, date_created, is_public, u
     return text
 
 
-def get_snippet_relation(relationset):
-    """
-    Get a text snippet for a :class:`.RelationSet` instance.
-    """
 
-    appellation_type = ContentType.objects.get_for_model(Appellation)
-    tokenizedContent = relationset.occursIn.tokenizedContent
-    annotated_words = []
-
-    # We'll use this to label highlighted tokens with their interpretations.
-    annotation_map = {}
-
-    fields = [
-        'source_content_type_id',
-        'object_content_type_id',
-        'source_object_id',
-        'object_object_id',
-        'predicate__tokenIds',
-        'predicate__interpretation__label',
-        'predicate__interpretation_id',
-    ]
-
-    appellation_ids = set()
-
-    # Pull out all Appellations that have a specific textual basis.
-    for relation in relationset.constituents.values(*fields):
-        for part in ['source', 'object']:
-            if relation.get('%s_content_type_id' % part, None) == appellation_type.id:
-                appellation_ids.add(relation['%s_object_id' % part])
-
-        # Predicates too, since we're interested in evidence for the relation.
-        if relation['predicate__tokenIds']:
-            tokenIds = relation['predicate__tokenIds'].split(',')
-            annotated_words.append(tokenIds)
-            for t in tokenIds:
-                annotation_map[t] = relation['predicate__interpretation__label']
-
-    appellation_fields = [
-        'tokenIds',
-        'interpretation__label',
-        'interpretation_id',
-        'interpretation__merged_with_id',
-        'interpretation__merged_with__label',
-    ]
-    for appellation in Appellation.objects.filter(pk__in=appellation_ids).values(*appellation_fields):
-
-        tokenIds = appellation['tokenIds'].split(',')
-        annotated_words.append(tokenIds)
-        for t in tokenIds:
-            if appellation['interpretation__merged_with_id']:
-                annotation_map[t] = appellation['interpretation__merged_with__label']
-            else:
-                annotation_map[t] = appellation['interpretation__label']
-
-    # Group sequences of tokens from appellations together if they are in close
-    #  proximity.
-    grouped_words = []
-    for tokenSeq in sorted(annotated_words, key=lambda s: min([int(t) for t in s])):
-        tokenSeqI = [int(t) for t in tokenSeq]
-        match = False
-        for i, tokensSeqGrouped in enumerate(grouped_words):
-            tokensSeqGroupedI = [int(t) for t in tokensSeqGrouped]
-            # If the current token sequence is contained within or overlaps with
-            #  the token sequence group, then add it to the current group and
-            #  exit.
-            if (min(tokenSeqI) >= min(tokensSeqGroupedI) - 10 and max(tokenSeq) <= max(tokensSeqGroupedI) + 10) or \
-               (min(tokenSeqI) - 10 <= min(tokensSeqGroupedI) <= max(tokenSeqI) + 10) or \
-               (max(tokenSeqI) + 10 >= max(tokensSeqGroupedI) >= min(tokenSeqI) - 10):
-               grouped_words[i] = tokensSeqGrouped + tokenSeq
-               match = True
-               break
-
-        if not match:    # Sequence belongs to its own group (for now).
-            grouped_words.append(tokenSeq)
-
-    # Now build the snippet.
-    combined_snippet = u""
-    for tokenSeq in grouped_words:
-        snippet = u""
-        start_index = max(0, min([int(t) for t in tokenSeq])) - 5
-        end_index = max([int(t) for t in tokenSeq]) + 5
-        for i in range(start_index, end_index):
-            match = re.search(r'<word id="'+str(i)+'">([^<]*)</word>', tokenizedContent, re.M|re.I)
-            if not match:
-                continue
-            word = ""
-            if str(i) in tokenSeq:
-                # Tooltip shows the interpretation (Concept) for this
-                #  Appellation.
-                word = u"<strong data-toggle='tooltip' title='%s' class='text-warning text-snippet'>%s</strong>" % (annotation_map[str(i)], match.group(1))
-            else:
-                word = match.group(1)
-            snippet = u'%s %s' % (snippet, word)
-        combined_snippet += u' ...%s... ' % snippet.strip()
-    return SafeText(combined_snippet)
-
-
-def get_snippet(appellation):
-    """
-    Extract the text content surrounding (and including) an
-    :class:`.Appellation` instance.
-
-    Parameters
-    ----------
-    appellation : :class:`.Appellation`
-
-    Returns
-    -------
-    snippet : :class:`django.utils.safestring.SafeText`
-        Includes emphasis tags surrounding the :class:`.Appellation`\'s
-        tokens.
-    """
-    if not appellation['tokenIds']:
-        return SafeText('No snippet is available for this appellation')
-
-    tokenizedContent = appellation['occursIn__tokenizedContent']
-    annotated_words = [i.strip() for i in appellation['tokenIds'].split(',')]
-    middle_index = int(annotated_words[max(len(annotated_words)/2, 0)])
-    start_index = max(middle_index - 10, 0)
-    end_index = middle_index + 10
-    snippet = ""
-    for i in range(start_index, end_index):
-        match = re.search(r'<word id="'+str(i)+'">([^<]*)</word>', tokenizedContent, re.M|re.I)
-        if not match:
-            continue
-
-        word = ""
-
-        if str(i) in annotated_words:
-            word = u"<strong class='text-warning text-snippet'>%s</strong>" % match.group(1)
-        else:
-            word = match.group(1)
-        snippet = u'%s %s' % (snippet, word)
-    return SafeText(u'...%s...' % snippet.strip())
 
 
 @shared_task
