+++ annotations/utils.py
@@ -2,6 +2,8 @@
 General-purpose helper functions.
 """
 
+from django.conf import settings
+
 import re
 
 
@@ -11,3 +13,24 @@ def help_text(text):
     fields when writing long help_texts.
     """
     return re.sub('(\s+)', ' ', text)
+
+
+def basepath(request):
+    """
+    Generate the base path (domain + path) for the site.
+
+    TODO: Do we need this anymore?
+
+    Parameters
+    ----------
+    request : :class:`django.http.request.HttpRequest`
+
+    Returns
+    -------
+    str
+    """
+    if request.is_secure():
+        scheme = 'https://'
+    else:
+        scheme = 'http://'
+    return scheme + request.get_host() + settings.SUBPATH
